Logging ToB banter output to wrong debug log
Most of the v21 banters (and a few prior) were logging to the SoA debug log. If the SoA banters weren't installed, the install
errors out. This has probably been here for a while, but never noticed, because who would install ToB crossmod and not SoA? I
only found it while testing ToB stuff.