Right, so a few changes here and there.
The last few versions of Crossmod have made it pretty clear that it desperately needed a re-think as to how it’s structured,
as it’s incredibly unmaintainable in its current form. Having given it a lot of thought, I’ve tried to implement a framework
with prioritization on making it easier for modders to add content via Crossmod while making the whole thing sustainable
going forward.

To that end, much of the content has been automated. Just drop content into the right folders (a d or tpa file or both) with
an associated tra file and Crossmod does the rest. If it’s a new mod to Crossmod, there’s one additional step to tell
Crossmod how to find your mod, but that’s it. It’s also fairly open-ended, so Crossmod can be used to have NPCs react to
quest mods (which it already does in a few places), or even quest mods with one another. Naturally this doesn’t come without
tradeoffs, as the automation requires stricter naming conventions for files and folders.

Logging has also been reduced significantly in verbosity. Generating the high number of unique-but-tediously-similar
feedback that the installer echoed and logged was one of the main time-sinks of deploying new builds. Installer feedback has
been eliminated and, while the log files remain, they’re much-reduced and composed of auto-generated strings.

Speed is a concern, and one of the reasons why I’m committing this now is so that I can turn it out to some folks for
testing. On the one hand the new framework only does one file check per mod, total, over an installation whereas the old
methodology would do two file-checks for every bit of Crossmod—meaning that it would be redundantly checking for a file
multiple times even after the first check failed. On the flipside the new framework has redundancy in spades as it will
check every detected mod on its list against every other mod for content, whereas the old method had no such speculative
checking. I suspect that the new framework will be generally faster until a certain threshold of installed mods is breached,
as the cross-checks grow exponentially with the number of detected mods. The hope is that the threshold is sufficiently high
that players are not inconvenienced. If not, well, it’ll be time for another re-think.

One small feature that I’ve added is the ability to restrict partially-translated installs. For example, right now we have a
small handful of content translated into German, so if you install Crossmod with other mods that lack German strings,
Crossmod will install a mix of German and English content. There’s a hidden variable that will restrict non-English installs
to only install content that’s in their language. In future builds I’ll make this a little more accessible, depending on
interest.

The keen observer will note the use of %MOD_FOLDER% throughout the libraries, as a successful rollout of the framework here
will see it deployed to IWD Crossmod and elsewhere.

At any rate, while I’m confident we’ve got a stable framework there’s no way I renamed and re-ordered all these files
without introducing a bug *somewhere*. For now I’ll be focusing on speed testing and then, if all goes well, opening it up
to a broader audience for more robust testing.

Keep your eyes on the Crossmod forum for updates in the mean time. I’ll be providing a ‘how do I add content’ topic for
modders once I’m sure we won’t have to change something in the mod registration process.