Language handling
Various Crossmod bits rely on STATE_WHICH_SAYS to determine where to interject. I had hoped that %s would allow WeiDU to parse
the LANGUAGE tree of other mods, but I think it's built locally off the current mods' LANGUAGE tree. Whether it's that or
something else, the upshot is that it doesn't work. Instead we manually determine the installed language of these mods so we
know which folder to use for tra lookups. A future maintenance issue is that Crossmod will have to update if these mods add new
translations.